+++ celt/hybrid_tf_resolution_test.go
@@ -4,20 +4,20 @@ import "testing"
 
 func TestFillHybridTFResolution(t *testing.T) {
 	t.Run("weak transient", func(t *testing.T) {
-		tfRes := []int{9, 9, 9, 9}
+		tfRes := []int32{9, 9, 9, 9}
 		tfSelect := FillHybridTFResolution(tfRes, 3, true, true, true)
 		if tfSelect != 0 {
 			t.Fatalf("tfSelect = %d, want 0", tfSelect)
 		}
 		for i, want := range []int{1, 1, 1, 9} {
-			if tfRes[i] != want {
+			if tfRes[i] != int32(want) {
 				t.Fatalf("tfRes[%d] = %d, want %d", i, tfRes[i], want)
 			}
 		}
 	})
 
 	t.Run("low bitrate transient", func(t *testing.T) {
-		tfRes := []int{9, 9, 9}
+		tfRes := []int32{9, 9, 9}
 		tfSelect := FillHybridTFResolution(tfRes, len(tfRes), true, false, true)
 		if tfSelect != 1 {
 			t.Fatalf("tfSelect = %d, want 1", tfSelect)
@@ -30,7 +30,7 @@ func TestFillHybridTFResolution(t *testing.T) {
 	})
 
 	t.Run("default transient", func(t *testing.T) {
-		tfRes := []int{9, 9, 9}
+		tfRes := []int32{9, 9, 9}
 		tfSelect := FillHybridTFResolution(tfRes, len(tfRes), true, false, false)
 		if tfSelect != 0 {
 			t.Fatalf("tfSelect = %d, want 0", tfSelect)
